refactor card types
