--- src/model/CardType.ts
@@ -1,13 +1,0 @@
-import { ReactNode } from 'react'
-
-export interface CardField {
-  ID: string
-  name: string
-}
-
-export interface CardType {
-  ID: string
-  name: string
-  fields: CardField[]
-  templates: string[]
-}
